util: Disable capturing of subpatterns
It seems that the pattern we use will create a subpattern match for
every character of the value. This appears to then result in a stack
overflow when matching very long values.

It might be an idea to limit the overall length of an environment
variable that is uploaded. But, this should fix the issue of crashes.

Fixes a stack overflow caused by long environment variables.

See: https://github.com/GNOME/gnome-session/commit/3b57d117